Squashed commit of the following:
commit 07a0ae052ca78620a5642354cf7236db201e8d96
Author: SueDonham <Susan.Donham@pm.me>
Date:   Sun Oct 6 10:52:40 2024 -0500

    Fix: exit if run from wrong directory

commit 604fd618b39d5cbaa25eeee93da9c7b5c0577952
Author: SueDonham <Susan.Donham@pm.me>
Date:   Sun Oct 6 09:14:45 2024 -0500

    Tweak install scripts + readme

commit 9bbb81f52fec67e8204e42bba5d8afb97808fba8
Merge: 142e7b7c 57d290f1
Author: SueDonham <Susan.Donham@pm.me>
Date:   Sun Sep 29 08:16:48 2024 -0500

    Merge PR #4 for feature/hyprcursor-support

commit 57d290f1a562998cf0625ea60b11519a07c16d32
Author: Greentea <vckuznetsov@gmail.com>
Date:   Thu Sep 12 22:32:51 2024 +0300

    Added hyprcursor options to install.sh script

    Added `XCURSOR`, `HYPRCURSOR `and `DRY_RUN` environment variables
    `XCURSOR`=1 installs xcursor theme
    `HYPRCURSOR`=1 installss hyprcursor theme
    `DRY_RUN`=1 sets dry run mode (no real changes will be made)

commit dd560511db55fd83b34052348620b0a2b0309ce5
Author: Greentea <vckuznetsov@gmail.com>
Date:   Thu Sep 12 21:15:55 2024 +0300

    Readd hyprcursor theme support

    `cursors/build-hl.sh` to build themes
    `cursors/install-hl.sh` to install themes to user's icons path
    Prebuilt themes are included in `hl-dark` and `hl-light`
    Ready-to-use themes are included in `dist-hl-dark` and `dist-hl-light`

+++ cursors/install.sh
@@ -11,27 +11,41 @@ else
   DEST_DIR="$HOME/.icons"
 fi
 
+# Ensure user is in the right dir:
+SCRIPT_DIR=$( cd -- "$( dirname -- "${BASH_SOURCE[0]}" )" &> /dev/null && pwd )
+[[ "$PWD" != "$SCRIPT_DIR" ]] && cd "$SCRIPT_DIR"
 
+
+# Select which theme(s) to install:
+read -p $'Install Xcursor, hyprcursor, or both?\n1) Xcursor\n2) hyprcursor\n3) Both\nEnter selection [1|2|3]: ' format
+case $format in
+	1) XCURSOR=true ;;
+	2) HYPRCURSOR=true ;;
+	3) XCURSOR=true && HYPRCURSOR=true ;;
+	*) echo "Invalid selection; exiting" && exit ;;
+esac
 
 
 # Copy light or dark cursor variant to DEST_DIR:
-copy_variant(){
-var=$1
-dest_dir=$DEST_DIR/$THEME_NAME-cursors-$var
+copy_variant() {
+	var=$1
+	dest_dir="$DEST_DIR/$THEME_NAME-cursors-$var"
 
+	[[ -d "$dest_dir" ]] && rm -r "$dest_dir"
 
-if [ -d "$dest_dir" ]; then
-  rm -r "$dest_dir"
-fi
-
-mkdir -p $dest_dir
-cp -r dist-$var/* $dest_dir
+	mkdir -p $dest_dir
+	cp -r dist-$var/* $dest_dir
 }
 
 
-cd "$(dirname "$0")"
-copy_variant 'light'
-copy_variant 'dark'
+if [[ "$XCURSOR" ]] ; then
+  copy_variant 'light'
+  copy_variant 'dark'
+fi
+
+if [[ "$HYPRCURSOR" ]] ; then
+  copy_variant 'hl-light'
+  copy_variant 'hl-dark'
+fi
 
 echo "Cursors copied to $DEST_DIR"
-
